Move calendars to separate folder, use responses to update offer
